chore: create token balance endpoints file
